Add View, Controller, and AplMain classes for testing data insertion
- Created the View class to handle the user interface.
- Added the Controller class to manage the application logic.
- Introduced the AplMain class to start the testing application.
- Established a database connection and performed a test insertion.

Implementation of CRUD logic in PersonDAO

- Added functionality for creating, reading, updating, and deleting
records in the database within the PersonDAO class.
- Implemented the CRUD interface in all DAO classes for consistent data
management.

This commit adds the necessary classes and components to initiate a data
insertion test into the database and implements complete data management
functionality in the PersonDAO class. It lays a solid foundation for
future development and testing, ensuring data management consistency
throughout the project.